Fixup iteration when lru->q == glistn, not glist.
If the node being dequeued is the one currently being processed
by lru_thread, then the value of glistn is stable and is the next
queue entry as desired.

However if the node being dequeued is that pointed to by glistn,
then lru_thread will proximately fault, so glistn must be updated
to point to the next entry in the queue.

Also it is useless to update glist, since it already points to the
entry before glistn (obviously).

Signed-off-by: Matt Benjamin <matt@linuxbox.com>